move search like functions into own module, debounce of 250
