+++ app/src/js/main.js
@@ -1,10 +1,12 @@
 import formCtrl from './formCtrl';
 import fileUpload from './file-upload';
 import validation from './validation';
+import search from './search';
 import { checkSupport } from './v/index';
 
 checkSupport();
 
 formCtrl.init();
 fileUpload.init();
+search.init();
 validation.init();
